Fix calendar bug on home page and use async await for calendar API calls (#111)
The home page use to not show upcoming events because the function call was missing the `calendarType` param. This has been fixed.

Along the way, I've refactored the code a bit to use async/await as well as functional components for Home, Events & Wise pages.

No visual changes were made.